disabling previsualisation and warning for links
The homepage link has ben updated to support a warning if the link is only available in one langauge (can be specified in a new metadata field.  A new metadata field to disable the link for previsualisation is also implemented. CAUTION : there are two field in the metadata : homepage and program_page_url.  Those seems redundant and should be later deleted. HEADS-UP : you will STILL see the link for the preview even if it's disabled IF you have the rights to edit the dataset